requirements.txt
asgiref==3.2.7
Babel==2.8.0
boto3==1.12.37
botocore==1.15.37
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-crispy-forms==1.9.0
django-environ==0.4.5
django-friends==0.1.5
django-friendship==1.9.1
django-phonenumber-field==4.0.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
googlemaps==4.2.0
haversine==2.2.0
idna==2.9
jmespath==0.9.5
lxml==4.5.0
mysqlclient==1.4.6
numpy==1.18.2
phonenumbers==8.12.2
Pillow==7.0.0
pykml==0.2.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
